+++ server/controllers/posts.js
@@ -36,7 +36,7 @@ export const getPosts = async (req, res) => {
         .find({
           $and: [
             { creator: req.userId },
-            { groups: [] },
+            { originGroup: null },
             { editor: req.userId },
           ],
         })
@@ -71,6 +71,7 @@ export const createPost = async (req, res) => {
       groups: [q],
       creator: req.userId,
       createdAt: new Date().toISOString(),
+      originGroup: q,
     });
   } else {
     newPost = new postMessage({
@@ -78,11 +79,12 @@ export const createPost = async (req, res) => {
       creator: req.userId,
       createdAt: new Date().toISOString(),
       editor: [req.userId],
+      originGroup: null,
     });
   }
 
   try {
-    console.log(newPost);
+    // console.log(newPost);
     await newPost.save();
     const post = await postMessage
       .findById(newPost._id)
@@ -136,7 +138,7 @@ export const updatePost = async (req, res) => {
       })
       .populate("editor", "email")
       .populate("viewer", "email");
-    console.log(updatedPost);
+    // console.log(updatedPost);
     res.json(updatedPost);
   } catch (error) {
     console.log(error);
@@ -151,27 +153,48 @@ export const deletePost = async (req, res) => {
   try {
     if (q) {
       const grp = await group.findById(q);
-      if (
-        grp.access.includes(req.userId) &&
-        String(post.creator) === req.userId
-      ) {
-        if (post.groups.length > 1) {
-          await postMessage.findByIdAndUpdate(
-            _id,
-            { $pull: { groups: q } },
-            { new: true }
-          );
-          return res.status(200).json("Memory removed successfully");
+      if (post.groups.includes(q)) {
+        //requesting user is the creator of the post and he has access to CRUD the posts of the group
+        if (
+          grp.access.includes(req.userId) &&
+          String(post.creator) === req.userId
+        ) {
+          //If the post belongs to multiple groups
+          if (post.groups.length > 1) {
+            await postMessage.findByIdAndUpdate(
+              _id,
+              { $pull: { groups: q } },
+              { new: true }
+            );
+            return res.status(200).json("Memory removed successfully");
+          } else {
+            //post belongs to single group
+            if (post.originGroup) {
+              //check if the post is created in some group then delete the whole post
+              await postMessage.findByIdAndDelete(_id);
+              res.json({ message: "the data is successfully deleted" });
+            } else {
+              //else just update the post
+              await postMessage.findByIdAndUpdate(
+                _id,
+                { $pull: { groups: q } },
+                { new: true }
+              );
+              return res.status(200).json("Memory removed successfully");
+            }
+          }
         } else {
-          await postMessage.findByIdAndDelete(_id);
-          res.json({ message: "the data is successfully deleted" });
+          return res
+            .status(403)
+            .json({ message: "You don't have access to delete this post" });
         }
       } else {
         return res
-          .status(403)
-          .json({ message: "You don't have access to delete this post" });
+          .status(400)
+          .json({ message: "The Post does not belong to the specified group" });
       }
     } else {
+      //If the creator delets the post then it will be deleted parmanently
       if (String(post.creator) === req.userId) {
         await postMessage.deleteOne({ _id: _id });
         return res.status(200).json("The post was successfully deleted");
@@ -217,20 +240,20 @@ export const sharePost = async (req, res) => {
     return res.status(404).json({ message: "User not found" });
   }
   if (String(existingUser._id) === req.userId) {
-    return res.json({ message: "You can not share with yourself" });
+    return res.status(400).json({ message: "You can not share with yourself" });
   }
   try {
     const post = await postMessage.findById(_id);
     if (post.editor.includes(existingUser._id)) {
-      console.log("here");
-      return res.json({
+      // console.log("here");
+      return res.status(403).json({
         message: "The post id already shared with the user",
       });
     }
     if (shareData.access === 1) {
       let updatedPost;
       if (post.viewer.includes(existingUser._id)) {
-        console.log("here");
+        // console.log("here");
         updatedPost = await postMessage
           .findByIdAndUpdate(
             _id,
@@ -279,7 +302,7 @@ export const sharePost = async (req, res) => {
       }
     } else {
       if (post.viewer.includes(existingUser._id)) {
-        return res.json({
+        return res.status(400).json({
           message: "The post id already shared with the user",
         });
       } else {
@@ -310,6 +333,91 @@ export const sharePost = async (req, res) => {
     console.log(error);
   }
 };
+export const removeEditorAccess = async (req, res) => {
+  const { id: _id } = req.params;
+  const { memberId } = req.body;
+  if (
+    !mongoose.Types.ObjectId.isValid(_id) ||
+    !mongoose.Types.ObjectId.isValid(memberId)
+  ) {
+    return res.status(400).json({ message: "Invalid Post or Member" });
+  }
+  try {
+    const post = await postMessage.findOne({ _id: _id });
+    if (post.editor.includes(memberId) && String(post.creator) !== memberId) {
+      const update = await postMessage
+        .findByIdAndUpdate(
+          _id,
+          {
+            $pull: { editor: memberId },
+          },
+          {
+            new: true,
+          }
+        )
+        .populate({
+          path: "creator",
+          model: "user",
+          select: "name",
+        })
+        .populate({
+          path: "editDetails.editedBy",
+          model: "user",
+          select: "name",
+        })
+        .populate("editor", "email")
+        .populate("viewer", "email");
+      return res.status(200).json(update);
+    } else {
+      return res.status(400).json({ message: "Invalid Request" });
+    }
+  } catch (error) {
+    console.log(error);
+  }
+};
+export const removeFromeViewerAccess = async (req, res) => {
+  const { id: _id } = req.params;
+  const { memberId } = req.body;
+  if (
+    !mongoose.Types.ObjectId.isValid(_id) ||
+    !mongoose.Types.ObjectId.isValid(memberId)
+  ) {
+    return res.status(400).json({ message: "Invalid Post or Member" });
+  }
+  try {
+    const post = await postMessage.findOne({ _id: _id });
+    if (post.viewer.includes(memberId) && String(post.creator) !== memberId) {
+      const update = await postMessage
+        .findByIdAndUpdate(
+          _id,
+          {
+            $pull: { viewer: memberId },
+          },
+          {
+            new: true,
+          }
+        )
+        .populate({
+          path: "creator",
+          model: "user",
+          select: "name",
+        })
+        .populate({
+          path: "editDetails.editedBy",
+          model: "user",
+          select: "name",
+        })
+        .populate("editor", "email")
+        .populate("viewer", "email");
+      return res.status(200).json(update);
+    } else {
+      return res.status(400).json({ message: "Invalid Request" });
+    }
+  } catch (error) {
+    console.log(error);
+  }
+};
+
 export const likeGroupPost = async (req, res) => {
   const { groupid: q } = req.query;
   const { id: _id } = req.params;
@@ -371,5 +479,49 @@ export const likeGroupPost = async (req, res) => {
     }
   } catch (error) {
     console.log(error.message);
+  }
+};
+export const sharePostToGroup = async (req, res) => {
+  const { id: _id } = req.params;
+  const { groupId } = req.body;
+  if (
+    !mongoose.Types.ObjectId.isValid(_id) ||
+    !mongoose.Types.ObjectId.isValid(groupId)
+  ) {
+    return res.status(400).json({ message: "Invalid" });
+  }
+  try {
+    const post = await postMessage.findOne({ _id: _id });
+    if (post.groups.includes(groupId)) {
+      return res
+        .status(400)
+        .json({ message: "The post is already in that group" });
+    } else {
+      const update = await postMessage
+        .findByIdAndUpdate(
+          _id,
+          {
+            $push: { groups: groupId },
+          },
+          {
+            new: true,
+          }
+        )
+        .populate({
+          path: "creator",
+          model: "user",
+          select: "name",
+        })
+        .populate({
+          path: "editDetails.editedBy",
+          model: "user",
+          select: "name",
+        })
+        .populate("editor", "email")
+        .populate("viewer", "email");
+      return res.status(200).json(update);
+    }
+  } catch (error) {
+    console.log(error);
   }
 };
